requirements-clean.txt
# ── Core ──────────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# ── Database (Neon PostgreSQL) ────────────────────────────────────────────────
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
greenlet>=3.0.0

# ── Auth ──────────────────────────────────────────────────────────────────────
bcrypt>=4.1.0
PyJWT>=2.8.0

# ── ML ────────────────────────────────────────────────────────────────────────
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# ── External services (existing) ──────────────────────────────────────────────
openai>=1.10.0
composio>=0.10.0
requests>=2.31.0

# ── Utilities ─────────────────────────────────────────────────────────────────
httpx>=0.27.0
